removed unnecessary new file, broken tests and unnecessary else statement: created_at is always a valid Date
